@@ -3,36 +3,35 @@ using System;
 using System.Reactive.Linq;
 using TIKSN.Leveret.Interpretation.Abstractions;
 
-namespace TIKSN.Leveret.ViewModels
+namespace TIKSN.Leveret.ViewModels;
+
+public class MainWindowViewModel : ViewModelBase
 {
-    public class MainWindowViewModel : ViewModelBase
+    private readonly ObservableAsPropertyHelper<InterpretationResult> _executionResults;
+    private string _inputSourceCode;
+
+    public MainWindowViewModel(IInterpretationService interpretationService)
     {
-        private readonly ObservableAsPropertyHelper<InterpretationResult> _executionResults;
-        private string _inputSourceCode;
-
-        public MainWindowViewModel(IInterpretationService interpretationService)
+        if (interpretationService == null)
         {
-            if (interpretationService == null)
-            {
-                throw new ArgumentNullException(nameof(interpretationService));
-            }
-
-            _executionResults = this
-            .WhenAnyValue(x => x.InputSourceCode)
-            .Throttle(TimeSpan.FromMilliseconds(800))
-            .DistinctUntilChanged()
-            .Where(code => !string.IsNullOrWhiteSpace(code))
-            .SelectMany(interpretationService.InterpretationAsync)
-            .ObserveOn(RxApp.MainThreadScheduler)
-            .ToProperty(this, x => x.ExecutionResults);
+            throw new ArgumentNullException(nameof(interpretationService));
         }
 
-        public InterpretationResult ExecutionResults => _executionResults.Value;
+        _executionResults = this
+        .WhenAnyValue(x => x.InputSourceCode)
+        .Throttle(TimeSpan.FromMilliseconds(800))
+        .DistinctUntilChanged()
+        .Where(code => !string.IsNullOrWhiteSpace(code))
+        .SelectMany(interpretationService.InterpretationAsync)
+        .ObserveOn(RxApp.MainThreadScheduler)
+        .ToProperty(this, x => x.ExecutionResults);
+    }
 
-        public string InputSourceCode
-        {
-            get => _inputSourceCode;
-            set => this.RaiseAndSetIfChanged(ref _inputSourceCode, value);
-        }
+    public InterpretationResult ExecutionResults => _executionResults.Value;
+
+    public string InputSourceCode
+    {
+        get => _inputSourceCode;
+        set => this.RaiseAndSetIfChanged(ref _inputSourceCode, value);
     }
 }
